requirements.txt
asgiref==3.8.1
azure-core==1.31.0
azure-cosmos==4.7.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.4.0
django-tailwind==4.0.1
djongo==1.3.6
dnspython==2.6.1
idna==3.10
PyJWT==2.9.0
pymongo==4.8.0
pillow==10.2.0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
numpy==2.2.5

moviepy==2.1.2 
decorator==5.2.1
imageio==2.37.0 
imageio_ffmpeg==0.6.0 
numpy==2.2.5 
proglog==0.1.11 
tqdm==4.67.1

decorator==4.4.2 
moviepy==1.0.3